Zdjęcia profilowe (upload na serwer - nie działa na studentlive) Dodane komentarze

+++ GrupaEKA/Controllers/NewsController.cs
@@ -5,10 +5,10 @@ using System.Web;
 using System.Web.Mvc;
 using GrupaEka.Models;
 using System.Data;
+using System.Web.Security;
 
 namespace GrupaEka.Controllers
 {
-    [Authorize(Roles = "Admin")]
     public class NewsController : Controller
     {
         private GrupaEkaDB db = new GrupaEkaDB();
@@ -17,6 +17,7 @@ namespace GrupaEka.Controllers
 
         //
         // GET: /News/
+        [Authorize(Roles="admin")]
         public ActionResult Index()
         {
             var news = db.News.OrderByDescending(n => n.Date);
@@ -26,6 +27,7 @@ namespace GrupaEka.Controllers
 
         //
         // GET: /News/Create
+        [Authorize]
         public ActionResult Create()
         {
             ViewBag.Now = DateTime.Now;
@@ -38,6 +40,7 @@ namespace GrupaEka.Controllers
 
         //
         // POST: /News/Create
+        [Authorize]
         [HttpPost]
         public ActionResult Create(NewsTimeViewModel newNewsTime)
         {
@@ -49,15 +52,19 @@ namespace GrupaEka.Controllers
                 Int32.TryParse(newNewsTime.Minutes,out minutes);
                 TimeSpan time = new TimeSpan(hour, minutes, 0);
                 newNews.Date = newNews.Date + time;
-                var cats = db.NewsCategories.Where(r => newNewsTime.CategoryIDs.Contains(r.ID));                
-                foreach (var cat in cats)
+                try
                 {
-                    newNews.NewsCategories.Add(cat);                    
+                    var cats = db.NewsCategories.Where(r => newNewsTime.CategoryIDs.Contains(r.ID)).ToList();
+                    foreach (var cat in cats)
+                    {
+                        newNews.NewsCategories.Add(cat);
+                    }
                 }
+                catch (NotSupportedException) { /* 0 categories */ }
                 db.News.Add(newNews);
                 db.SaveChanges();
 
-                return RedirectToAction("Index");
+                return RedirectToAction("Index","Home");
             }
             else
             {
@@ -73,18 +80,21 @@ namespace GrupaEka.Controllers
         {
             News news = db.News.Find(id);
             if (news == null)
-                return RedirectToAction("Index");
+                return RedirectToAction("Index", "Home");
             return View("Details", news);
         }
 
         // 
         // GET: /News/Edit
+        [Authorize]
         public ActionResult Edit(int id)
         {
             ViewBag.NewsCategories = db.NewsCategories;
             News news = db.News.Find(id);
             if (news == null)
-                return RedirectToAction("Index");
+                return RedirectToAction("Index", "Home");
+            if (!news.Author.Equals(Membership.GetUser().UserName) && !Roles.IsUserInRole(Membership.GetUser().UserName, "admin"))
+                return RedirectToAction("Details", new { ID = news.ID });
             NewsTimeViewModel editNews = new NewsTimeViewModel();
             editNews.News = news;
             editNews.Hour = news.Date.Hour.ToString();
@@ -95,12 +105,17 @@ namespace GrupaEka.Controllers
 
         // 
         // POST: /News/Edit 
+        [Authorize]
         [HttpPost]
         public ActionResult Edit(NewsTimeViewModel model)
         {
             try
             {
                 var news = db.News.Find(model.News.ID);
+                if (news == null)
+                    return RedirectToAction("Index", "Home");
+                if (!news.Author.Equals(Membership.GetUser().UserName) && !Roles.IsUserInRole(Membership.GetUser().UserName, "admin"))
+                    return RedirectToAction("Details", new { ID = news.ID });
                 int hour, minutes;
                 Int32.TryParse(model.Hour, out hour);
                 Int32.TryParse(model.Minutes, out minutes);
@@ -108,12 +123,16 @@ namespace GrupaEka.Controllers
                 news.Date = model.News.Date + time;
                 news.Text = model.News.Text;
                 news.Title = model.News.Title;
-                var cats = db.NewsCategories.Where(r => model.CategoryIDs.Contains(r.ID));
-                news.NewsCategories.Clear();
-                foreach (var cat in cats)
+                try
                 {
-                    news.NewsCategories.Add(cat);
+                    var cats = db.NewsCategories.Where(r => model.CategoryIDs.Contains(r.ID));
+                    news.NewsCategories.Clear();
+                    foreach (var cat in cats)
+                    {
+                        news.NewsCategories.Add(cat);
+                    }
                 }
+                catch (NotSupportedException) { /* 0 categories */ }
                 UpdateModel(news);
                 db.SaveChanges();
                 return RedirectToAction("Details", new { id = model.News.ID });
@@ -129,24 +148,42 @@ namespace GrupaEka.Controllers
 
         // 
         // GET: /News/Delete
+        [Authorize]
         public ActionResult Delete(int id)
         {
             News news = db.News.Find(id);
             if (news == null)
-                return RedirectToAction("Index");
+                return RedirectToAction("Index", "Home");
+            if (!news.Author.Equals(Membership.GetUser().UserName) && !Roles.IsUserInRole(Membership.GetUser().UserName, "admin"))
+                return RedirectToAction("Details", new { ID = news.ID });
             return View(news);
         }
 
         // 
         // POST: /News/Delete 
+        [Authorize]
         [HttpPost]
         public RedirectToRouteResult Delete(int id, FormCollection collection)
         {
             News news = db.News.Find(id);
+
+            if (news == null)
+                return RedirectToAction("Index", "Home");
+            if (!news.Author.Equals(Membership.GetUser().UserName) && !Roles.IsUserInRole(Membership.GetUser().UserName, "admin"))
+                return RedirectToAction("Details", new { ID = news.ID });
+
+            //delete comments for news - unnecessary when foreign key's Delete Rule=Cascade
+            //List<NewsComment> nc = db.NewsComments.Where(c => c.News.ID == id).ToList();
+            //foreach(var c in nc)
+            //{
+            //    db.NewsComments.Remove(c);
+            //}
+            //end of delete comments
+
             db.News.Remove(news);
             db.SaveChanges();
 
-            return RedirectToAction("Index");
+            return RedirectToAction("Index", "Home");
         }
 
         #endregion News
@@ -164,6 +201,7 @@ namespace GrupaEka.Controllers
 
         //
         // GET: /News/CategoryCreate
+        [Authorize]
         public ActionResult CategoryCreate()
         {            
             return View();
@@ -171,6 +209,7 @@ namespace GrupaEka.Controllers
 
         //
         // POST: /News/CategoryCreate
+        [Authorize]
         [HttpPost]
         public ActionResult CategoryCreate(NewsCategory category)
         {
@@ -193,6 +232,7 @@ namespace GrupaEka.Controllers
 
         // 
         // GET: /News/CategoryDetails
+        [Authorize]
         public ActionResult CategoryDetails(int id)
         {
             NewsCategory category = db.NewsCategories.Find(id);
@@ -203,6 +243,7 @@ namespace GrupaEka.Controllers
 
         // 
         // GET: /News/CategoryEdit
+        [Authorize]
         public ActionResult CategoryEdit(int id)
         {
             NewsCategory category = db.NewsCategories.Find(id);
@@ -213,6 +254,7 @@ namespace GrupaEka.Controllers
 
         // 
         // POST: /News/CategoryEdit 
+        [Authorize]
         [HttpPost]
         public ActionResult CategoryEdit(NewsCategory newscategory)
         {
@@ -227,6 +269,7 @@ namespace GrupaEka.Controllers
 
         // 
         // GET: /News/CategoryDelete
+        [Authorize]
         public ActionResult CategoryDelete(int id)
         {
             NewsCategory category = db.NewsCategories.Find(id);
@@ -237,6 +280,7 @@ namespace GrupaEka.Controllers
 
         // 
         // POST: /News/CategoryDelete 
+        [Authorize]
         [HttpPost]
         public RedirectToRouteResult CategoryDelete(int id, FormCollection collection)
         {
@@ -248,5 +292,102 @@ namespace GrupaEka.Controllers
         }
 
         #endregion NewsCategory
+
+        #region NewsComment
+        
+        //
+        // POST: /News/CommentCreate
+        [Authorize]
+        [HttpPost]
+        public RedirectToRouteResult CommentCreate(NewsCommentViewModel model)
+        {
+            NewsComment comment = new NewsComment();
+            comment.Author = Membership.GetUser().UserName;
+            comment.Date = DateTime.Now;
+            comment.Text = model.Text;
+            if (ModelState.IsValid)
+            {
+                //int NewsID = Int32.Parse(values["NewsID"]);
+                var news = db.News.Find(model.NewsID);
+                news.NewsComments.Add(comment);
+                db.SaveChanges();
+            }
+            else
+            {
+                ViewBag.CommentError = "Błąd! Komentarza nie dodano.";
+            }
+            return RedirectToAction("Details", new { ID = model.NewsID });
+        }
+
+        //
+        // GET: /News/CommentEdit
+        [Authorize]
+        public ActionResult CommentEdit(int id)
+        {
+            var newsComment = db.NewsComments.Find(id);
+            if (newsComment == null)
+                return RedirectToAction("Index", "Home");
+            if (!newsComment.Author.Equals(Membership.GetUser().UserName) && !Roles.IsUserInRole(Membership.GetUser().UserName,"admin"))
+                return RedirectToAction("Details", new { ID = newsComment.News.ID });
+            return View(newsComment);
+        }
+
+        //
+        // POST: /News/CommentEdit
+        [Authorize]
+        [HttpPost]
+        public ActionResult CommentEdit(FormCollection values)
+        {
+            int CommentID = Int32.Parse(values["ID"]);
+            var newsComment = db.NewsComments.Find(CommentID);
+            if(newsComment == null)
+                return RedirectToAction("Index", "Home");
+            if (!newsComment.Author.Equals(Membership.GetUser().UserName) && !Roles.IsUserInRole(Membership.GetUser().UserName, "admin"))
+                return RedirectToAction("Details", new { ID = newsComment.News.ID });
+            try
+            {   
+                newsComment.Date = DateTime.Now;
+                newsComment.Text = values["Text"];
+                UpdateModel(newsComment);
+                db.SaveChanges();   
+            }
+            catch (Exception)
+            {
+                ModelState.AddModelError("", "Błąd edycji!");
+            }
+            return RedirectToAction("Details", new { ID = newsComment.News.ID }); 
+        }
+
+        // 
+        // GET: /News/CommentDelete
+        [Authorize]
+        public ActionResult CommentDelete(int id)
+        {
+            var newsComment = db.NewsComments.Find(id);
+            if (newsComment == null)
+                return RedirectToAction("Index", "Home");
+            if(!newsComment.Author.Equals(Membership.GetUser().UserName) && !Roles.IsUserInRole(Membership.GetUser().UserName, "admin"))
+                return RedirectToAction("Details", new { ID = newsComment.News.ID });
+            return View(newsComment);
+        }
+
+        // 
+        // POST: /News/CommentDelete 
+        [Authorize]
+        [HttpPost]
+        public RedirectToRouteResult CommentDelete(int id, FormCollection collection)
+        {
+            var newsComment = db.NewsComments.Find(id);
+            if (newsComment == null)
+                return RedirectToAction("Index", "Home");
+            if(!newsComment.Author.Equals(Membership.GetUser().UserName) && !Roles.IsUserInRole(Membership.GetUser().UserName, "admin"))
+                return RedirectToAction("Details", new { ID = newsComment.News.ID });                
+            db.NewsComments.Remove(newsComment);
+            db.SaveChanges();
+
+            return RedirectToAction("Details", new { ID = newsComment.News_ID });
+        }
+
+        #endregion NewsComment
     }
 }
